requirements.txt
accelerate
arviz==0.17.1
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.3.0
cloudpickle==3.0.0
cons==0.4.6
contourpy==1.1.1
cramjam==2.8.2
cycler==0.12.0
datasets
docopt==0.6.2
duckdb==0.9.0
etuples==0.3.9
fastparquet==2023.8.0
fastprogress==1.0.3
filelock==3.12.4
fonttools==4.43.0
fsspec==2024.3.1
h5netcdf==1.3.0
h5py==3.10.0
idna==3.4
importlib_resources==6.3.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.42.0
logical-unification==0.4.6
MarkupSafe==2.1.3
matplotlib==3.8.0
medspacy
miniKanren==1.0.3
mpmath==1.3.0
multipledispatch==1.0.0
nestly==0.6.1
networkx==3.1
numba==0.59.1
numpy==1.24.4
openai
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pipreqs==0.4.13
psutil==5.9.5
pyarrow==13.0.0
pymc==5.11.0
pymc_bart==0.5.11
pyparsing==3.1.1
pytensor==2.18.6
python-dateutil==2.8.2
python-dotenv
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.12.0
seaborn
SCons==4.5.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
toolz==0.12.1
torch==2.0.1
transformers
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
xarray==2024.2.0
xarray-einstats==0.7.0
yarg==0.1.9
zipp==3.18.1
